requirements.txt

aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
Django==4.2
djangorestframework==3.14.0
frozenlist==1.3.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
keyboard==0.13.5
magic-filter==1.0.9
multidict==6.0.4
pytz==2022.7.1
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.14
yarl==1.8.2
